adding src/vtp/cli/_arguments.py
