+++ src/vtp/core/common.py
@@ -128,84 +128,6 @@ class Common:
         Common._configured = True
 
     @staticmethod
-    def cast_thing_to_list(argv):
-        """Primarly used by the argparse function in the operation classes"""
-        if isinstance(argv, dict):
-            new_argv = []
-            for key, value in argv.items():
-                if isinstance(value, bool):
-                    if value:
-                        new_argv.append("--" + key)
-                elif value is not None:
-                    new_argv.extend(["--" + key, str(value)])
-            return new_argv
-        return argv
-
-    # Tbe below are options that are shared across the various
-    # operations.  Options that are unique to one operation are
-    # located in that file.
-
-    @staticmethod
-    def add_blank_ballot(parser):
-        """Add blank_ballot option"""
-        parser.add_argument(
-            "--blank_ballot",
-            help="overrides an address - specifies the specific blank ballot",
-        )
-
-    @staticmethod
-    def add_election_data_dir(parser):
-        """Add election_data option"""
-        defval = "."
-        parser.add_argument(
-            "-e",
-            "--election_data_dir",
-            default=defval,
-            help=f"specify a absolute or relative path to the ElectionData tree (def={defval})",
-        )
-
-    @staticmethod
-    def add_merge_contests(parser):
-        """Add merge_contests option"""
-        parser.add_argument(
-            "-m",
-            "--merge_contests",
-            action="store_true",
-            help="Will immediately merge the ballot contests (to main)",
-        )
-
-    @staticmethod
-    def add_minimum_cast_cache(parser):
-        """Add minimum_cast_cache option"""
-        parser.add_argument(
-            "--minimum_cast_cache",
-            type=int,
-            default=100,
-            help="the minimum number of cast ballots required prior to merging (def=100)",
-        )
-
-    @staticmethod
-    def add_printonly(parser):
-        """Add printonly option"""
-        parser.add_argument(
-            "-n",
-            "--printonly",
-            action="store_true",
-            help="will printonly and not write to disk (def=True)",
-        )
-
-    @staticmethod
-    def add_verbosity(parser):
-        """Add verbosity option"""
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-
-    @staticmethod
     def verify_election_data_dir(election_data_dir: str):
         """
         Verify that election_data_dir is an existing directory.
